JMK_64: S3 implementation

+++ src/test/java/com/jobmatrix/dto/FreelancerDTOTest.java
@@ -9,6 +9,7 @@ import jakarta.validation.ValidatorFactory;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
 import java.util.Set;
 import java.util.UUID;
 
@@ -90,7 +91,7 @@ public class FreelancerDTOTest {
     // Negative Case: Negative hourly rate
     @Test
     public void testFreelancerDTO_HourlyRate_Negative() {
-        freelancerDTO.setHourlyRate(-10.0);
+        freelancerDTO.setHourlyRate(BigDecimal.valueOf(-10.0));
         Set<ConstraintViolation<FreelancerDTO>> violations = validator.validate(freelancerDTO);
         assertEquals(1, violations.size(), "Expected 1 validation error for negative hourly rate");
     }
